Fixed Search results RecyclerView issues with it overlapping the bottom buttons . Refactored SearchViewModel and removed it, integrated the items bindings inside the Main MapsViewModel
