+++ app/src/main/java/com/grandtour/ev/evgrandtour/ui/mainMapsView/models/SearchResultModel.java
@@ -1,8 +1,8 @@
-package com.grandtour.ev.evgrandtour.ui.mainMapsView.search;
+package com.grandtour.ev.evgrandtour.ui.mainMapsView.models;
 
 import android.support.annotation.NonNull;
 
-public class SearchResultViewModel {
+public class SearchResultModel {
 
     @NonNull
     public final Integer searchResultId;
@@ -13,7 +13,7 @@ public class SearchResultViewModel {
     @NonNull
     public final SearchViewResultClickListener clickListener;
 
-    public SearchResultViewModel(@NonNull Integer searchResultId, @NonNull String searchResultTourId, @NonNull String searchResultName,
+    public SearchResultModel(@NonNull Integer searchResultId, @NonNull String searchResultTourId, @NonNull String searchResultName,
             @NonNull SearchViewResultClickListener clickListener) {
         this.searchResultId = searchResultId;
         this.searchResultTourId = searchResultTourId;
